[#1212] Move data providers to designated class
